fix #407: add some composition of a large component
