+++ packages/frontapp/src/app/components/BookInfo/DescriptionBook.spec.tsx
@@ -1,0 +1,1 @@
+describe('render DescriptionBook component', () => {});
