+++ packages/frontapp/src/app/components/BookInfo/index.tsx
@@ -1,17 +1,12 @@
 import styled from '@emotion/styled';
-import bookImage from '../../../assets/MOC-data/BookImage.png';
-import React, { FC, useCallback, useEffect, useMemo, useState } from 'react';
+import React, { FC, useCallback, useEffect, useState } from 'react';
 import { colors, dimensions } from '@mimir/ui-kit';
 import { ReactComponent as Claim } from '../../../assets/claim.svg';
 import { ReactComponent as Edit } from '../../../assets/Edit.svg';
 import { ReactComponent as Remove } from '../../../assets/Remove.svg';
 import { ReactComponent as EnableNotifySvg } from '../../../assets/NoNotification.svg';
 import { ReactComponent as CancelNotifySvg } from '../../../assets/CancelNotification.svg';
-import {
-  GetMaterialByIdQueryResult,
-  Status,
-  useGetAllLocationsQuery,
-} from '@mimir/apollo-client';
+import { Status } from '@mimir/apollo-client';
 import { DateTime } from '@mimir/global-types';
 import Button from '../Button';
 import ClaimOperation from '../ClaimOperation';
@@ -22,7 +17,6 @@ import {
   periodOfKeeping,
 } from '../../models/helperFunctions/converTime';
 import SuccessMessage from '../SuccessMessage';
-import { listOfGenres } from '../../../assets/SearchConsts';
 import {
   GetAllTakenItemsDocument,
   GetMaterialByIdDocument,
@@ -39,13 +33,11 @@ import {
 import { useAppSelector } from '../../hooks/useTypedSelector';
 import ErrorMessage from '../ErrorMessge';
 import AskManagerForm from '../AskManagerForm';
-import { WrapperInput } from '../Search';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 import { RolesTypes } from '@mimir/global-types';
-import Dropdown, { IDropdownOption } from '../Dropdown';
-import BookStatus from '../BookStatus';
-import { toast } from 'react-toastify';
+import { IDropdownOption } from '../Dropdown';
 import { TUserLocation } from '../../store/slices/userSlice';
+import DescriptionBook from './DescriptionBook';
 
 export const BookHolder = styled.div`
   width: 100%;
@@ -60,16 +52,6 @@ export const BookHolder = styled.div`
   }
 `;
 
-export const BookImage = styled.img`
-  display: inline-block;
-  width: 12rem;
-  height: 19.5rem;
-  border-radius: 10px;
-  @media (max-width: ${dimensions.phone_width}) {
-    margin-right: ${dimensions.base};
-  }
-`;
-
 export const ShortDescriptionWrapper = styled.div`
   display: flex;
   justify-content: space-between;
@@ -80,54 +62,9 @@ export const ShortDescriptionWrapper = styled.div`
   }
 `;
 
-export const ShortDescription = styled.div`
-  width: 100%;
-  margin-left: ${dimensions.xl_2};
-  @media (max-width: ${dimensions.phone_width}) {
-    margin: 0;
-  }
-`;
-
-export const TitleBook = styled.h3`
-  font-weight: 700;
-  margin-bottom: ${dimensions.xl_2};
-  font-size: ${dimensions.xl};
-  line-height: ${dimensions.xl_2};
-  color: ${colors.main_black};
-  @media (max-width: ${dimensions.phone_width}) {
-    margin-bottom: 0;
-    margin-top: ${dimensions.base};
-  }
-`;
-
-export const Topic = styled.p`
-  margin: ${dimensions.base} 0 ${dimensions.xs_2} 0;
-  font-weight: 500;
-  font-size: ${dimensions.base};
-  line-height: ${dimensions.xl};
-  color: ${colors.main_black};
-`;
-
-export const TopicDescription = styled.p`
-  font-weight: 300;
-  font-size: ${dimensions.base};
-  line-height: ${dimensions.xl};
-  color: ${colors.main_black};
-`;
-
 export const LongDescription = styled.div`
   margin-top: ${dimensions.xl_2};
   grid-column: 1 / span 3;
-`;
-
-const OpenLink = styled.a`
-  cursor: pointer;
-  margin: ${dimensions.xs_2} 0;
-  font-weight: 300;
-  color: ${colors.accent_color};
-  font-size: ${dimensions.base};
-  line-height: ${dimensions.xl};
-  text-decoration: underline;
 `;
 
 export const Description = styled.p`
@@ -135,18 +72,6 @@ export const Description = styled.p`
   font-size: ${dimensions.base};
   line-height: ${dimensions.xl};
   color: ${colors.main_black};
-`;
-
-const StyledStatus = styled.div`
-  font-size: ${dimensions.base};
-`;
-
-export const WrapperInfo = styled.div`
-  display: flex;
-  @media (max-width: ${dimensions.phone_width}) {
-    flex-direction: column;
-    align-items: center;
-  }
 `;
 
 export const WrapperButtons = styled.div`
@@ -163,45 +88,12 @@ const StyledButton = styled(Button)`
   width: 100%;
 `;
 
-const StyledInput = styled.input`
-  width: 19rem;
-  border: none;
-  outline: none;
-  margin-left: ${dimensions.xs_2};
-  color: ${colors.main_black};
-  margin-right: 0.12rem;
-`;
-const StyledInputDeadline = styled.input`
-  border: 0.5px solid #bdbdbd;
-  border-radius: ${dimensions.xl_3};
-  padding: 10px 0;
-  width: 3.7rem;
-  outline: none;
-  padding-left: ${dimensions.xl};
-  background: ${colors.bg_secondary};
-  :hover {
-    border: 0.5px solid ${colors.accent_color};
-  }
-  :focus {
-    border: 0.5px solid ${colors.accent_color};
-  }
-  @media (max-width: ${dimensions.tablet_width}) {
-    width: 100%;
-  }
-  @media (max-width: ${dimensions.phone_width}) {
-    width: 70%;
-  }
-  ::-webkit-inner-spin-button,
-  ::-webkit-outer-spin-button {
-    -webkit-appearance: none;
-    margin: 0;
-  }
-`;
-const TitleHolder = styled.p`
+export const TitleHolder = styled.p`
   font-weight: 700;
   font-size: ${dimensions.base};
   margin-bottom: ${dimensions.xs};
 `;
+
 const StyledTextArea = styled.textarea`
   border: none;
   outline: none;
@@ -240,23 +132,36 @@ const TextAreaWrapper = styled.div`
   }
 `;
 
-const RestyledDropdown = styled(Dropdown)`
-  margin-top: ${dimensions.xs_2};
-  max-width: 21.5rem;
-  width: 100%;
+export const OpenLink = styled.a`
+  cursor: pointer;
+  margin: ${dimensions.xs_2} 0;
+  font-weight: 300;
+  color: ${colors.accent_color};
+  font-size: ${dimensions.base};
+  line-height: ${dimensions.xl};
+  text-decoration: underline;
 `;
 
-const WrapperDropDown = styled.div`
-  margin-top: ${dimensions.base};
-  width: 21.5rem;
+export const Topic = styled.p`
+  margin: ${dimensions.base} 0 ${dimensions.xs_2} 0;
+  font-weight: 500;
+  font-size: ${dimensions.base};
+  line-height: ${dimensions.xl};
+  color: ${colors.main_black};
 `;
 
 type StatusType = Pick<Status, 'id' | 'person_id' | 'created_at' | 'status'>;
-type Location = {
+export type Location = {
   __typename?: 'Location' | undefined;
   id: string;
   location: string;
 };
+
+export interface INewData {
+  newTitle: string;
+  newAuthor: string;
+}
+
 export interface IBookInfoProps {
   person_id: number | undefined;
   src: string | null | undefined;
@@ -289,13 +194,6 @@ const BookInfo: FC<IBookInfoProps> = ({
   location,
 }) => {
   const { id, userRole } = useAppSelector((state) => state.user);
-  const {
-    data: allLocations,
-    error: errorLocations,
-    loading: loadingLocations,
-  } = useGetAllLocationsQuery({
-    skip: userRole === RolesTypes.READER,
-  });
 
   const { data: getNotificationsByPersonData } =
     useGetNotificationsByPersonQuery({
@@ -326,37 +224,24 @@ const BookInfo: FC<IBookInfoProps> = ({
     useState<boolean>(false);
   const [valueIsISBN, setValueIsISBN] = useState<string>('');
   const [editing, setEditing] = useState<boolean>(false);
-  const [newTitle, setNewTitle] = useState(title);
-  const [newAuthor, setNewAuthor] = useState(author);
+
   const [newCategory, setNewCategory] = useState(category);
   const [newLocation, setNewLocation] = useState<Location>(location);
-  const [newDescription, setNewDescription] = useState(
-    description
-      ? description
-      : ' Lorem ipsum dolor sit amet consectetur, adipisicing elit. Sequi impedit aliquid alias consequuntur! Totam sequi expedita sunt dolor obcaecati, iusto ducimus? Beatae ea, commodi ab repellat, corporis atque quasi, tempore sunt modi similique soluta nemo hic necessitatibus esse accusantium omnis neque rerum. Placeat tempore, fugiat unde consequuntur dolor tempora ducimus.'
-  );
   const [newDeadline, setNewDeadline] = useState(periodOfKeeping);
+  const [newDescriptionData, setNewDescriptionData] = useState<INewData>({
+    newAuthor: author,
+    newTitle: title,
+  });
+
+  const [newDescription, setNewDescription] = useState(description || '');
   const [deleteWarning, setDeleteWarning] = useState(false);
   const [isMaterialTakenByCurrentUser, setIsMaterialTakenByCurrentUser] =
     useState(false);
 
-  const handleChangeTitle = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setNewTitle(e.target.value);
-  };
-  const handleChangeAuthor = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setNewAuthor(e.target.value);
-  };
-  const handleChangeDescription = (e: any) => {
+  const handleChangeDescription = (
+    e: React.ChangeEvent<HTMLTextAreaElement>
+  ) => {
     setNewDescription(e.target.value);
-  };
-  const handleChangeDeadline = (e: any) => {
-    setNewDeadline(e.target.value);
-    e.target.value > 31 && setNewDeadline(31);
-    e.target.value <= 0 && setNewDeadline(1);
-  };
-
-  const handleChangeCategory = (option: IDropdownOption) => {
-    setNewCategory(option.value);
   };
 
   const [claimBook, { data }] = useClaimBookMutation({
@@ -375,18 +260,49 @@ const BookInfo: FC<IBookInfoProps> = ({
     refetchQueries: [GetMaterialByIdDocument, GetAllTakenItemsDocument],
   });
   const currentStatus = getStatus(statusInfo?.status, created_at);
+
   const dateConditionOfClaiming =
     data?.claimBook.__typename === 'Status' ? data.claimBook.created_at : null;
+
   const dateConditionOfExtending =
     infoOfProlong?.prolongClaimPeriod.__typename === 'Status'
       ? infoOfProlong?.prolongClaimPeriod.created_at
       : null;
+
   const errorConditionOfClaiming =
     data?.claimBook.__typename === 'Error' ? data.claimBook.message : null;
   const errorConditionOfExtending =
     infoOfProlong?.prolongClaimPeriod.__typename === 'Error'
       ? infoOfProlong?.prolongClaimPeriod.message
       : null;
+
+  const handleChangeNewDescriptionData = (
+    e: React.ChangeEvent<HTMLInputElement>
+  ) => {
+    setNewDescriptionData({
+      ...newDescriptionData,
+      [e.target.name]: e.target.value,
+    });
+  };
+
+  const handleChangeCategory = (option: IDropdownOption) => {
+    setNewCategory(option.value);
+  };
+
+  const handleChangeLocation = (option: TUserLocation) => {
+    const newLocation = {
+      id: option.id,
+      location: option.value,
+    };
+    setNewLocation(newLocation);
+  };
+
+  const handleChangeDeadline = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setNewDeadline(Number(e.target.value));
+    Number(e.target.value) > 31 && setNewDeadline(31);
+    Number(e.target.value) <= 0 && setNewDeadline(1);
+  };
+
   const claim = async () => {
     await claimBook({
       variables: {
@@ -394,7 +310,6 @@ const BookInfo: FC<IBookInfoProps> = ({
         identifier: valueIsISBN,
       },
     });
-
     setValueIsISBN('');
     setIsShowClaimModal(false);
     setIsMaterialTakenByCurrentUser(true);
@@ -411,17 +326,6 @@ const BookInfo: FC<IBookInfoProps> = ({
     setIsSuccessReturn(true);
   };
 
-  const currentLocationIndex = useMemo(
-    () =>
-      allLocations?.getAllLocations.findIndex((loc) => loc!.id === location.id),
-    [allLocations]
-  );
-
-  const currentGenreIndex = useMemo(
-    () => listOfGenres.findIndex((genre) => genre.value === category),
-    [category, listOfGenres]
-  );
-
   const prolongPeriod = async () => {
     await prolongTime({
       variables: {
@@ -437,8 +341,8 @@ const BookInfo: FC<IBookInfoProps> = ({
         identifier: identifier,
         type: type,
         location_id: Number(newLocation.id),
-        title: newTitle,
-        author: newAuthor,
+        title: newDescriptionData.newTitle,
+        author: newDescriptionData.newAuthor,
         category: newCategory,
         updated_at: getDates(updated_at).currentDate,
       },
@@ -461,16 +365,15 @@ const BookInfo: FC<IBookInfoProps> = ({
   const handleDeleteBtn = () => setDeleteWarning(true);
 
   const discardChanges = () => {
-    setNewTitle(title);
-    setNewAuthor(author);
+    setNewDescriptionData({
+      newAuthor: author,
+      newTitle: title,
+    });
     setNewCategory(category);
-    setNewDescription(
-      description
-        ? description
-        : ' Lorem ipsum dolor sit amet consectetur, adipisicing elit. Sequi impedit aliquid alias consequuntur! Totam sequi expedita sunt dolor obcaecati, iusto ducimus? Beatae ea, commodi ab repellat, corporis atque quasi, tempore sunt modi similique soluta nemo hic necessitatibus esse accusantium omnis neque rerum. Placeat tempore, fugiat unde consequuntur dolor tempora ducimus.'
-    );
+    setNewDescription(description ? description : ' ');
     setEditing(false);
   };
+
   useEffect(() => {
     if (searchParams.has('claimModal') && currentStatus === 'Free') {
       const claimModal = searchParams.get('claimModal')!;
@@ -478,12 +381,6 @@ const BookInfo: FC<IBookInfoProps> = ({
       setValueIsISBN(claimModal);
     }
   }, []);
-
-  useEffect(() => {
-    if (errorLocations) {
-      toast.error(errorLocations.message);
-    }
-  }, [errorLocations]);
 
   useEffect(() => {
     if (infoOfProlong?.prolongClaimPeriod.__typename === 'Status') {
@@ -525,14 +422,6 @@ const BookInfo: FC<IBookInfoProps> = ({
     setIsShowClaimModal(true);
   }, []);
 
-  const handleChangeDropdown = (option: TUserLocation) => {
-    const newLocation = {
-      id: option.id,
-      location: option.value,
-    };
-    setNewLocation(newLocation);
-  };
-
   const handleEnableNotifyButton = async () => {
     await createNotificationMutation({
       variables: {
@@ -561,109 +450,22 @@ const BookInfo: FC<IBookInfoProps> = ({
     <>
       <BookHolder>
         <ShortDescriptionWrapper>
-          <WrapperInfo>
-            <BookImage src={src || bookImage} />
-            <ShortDescription>
-              {editing ? (
-                <>
-                  <TitleHolder>Name </TitleHolder>
-                  <WrapperInput>
-                    <StyledInput
-                      type="text"
-                      value={newTitle}
-                      onChange={handleChangeTitle}
-                    />
-                  </WrapperInput>
-                </>
-              ) : (
-                <TitleBook>{title || 'Book Title'}</TitleBook>
-              )}
-              {!editing && <Topic>Genre: </Topic>}
-              {userRole === RolesTypes.READER ? (
-                <OpenLink>{category || 'Genres of book'}</OpenLink>
-              ) : editing ? (
-                <WrapperDropDown>
-                  <TitleHolder>Genre </TitleHolder>
-                  <RestyledDropdown
-                    options={listOfGenres}
-                    initIndex={currentGenreIndex}
-                    onChange={handleChangeCategory}
-                    placeholder="Enter genre"
-                  />
-                </WrapperDropDown>
-              ) : (
-                <TopicDescription>
-                  {category || 'Genres of book'}
-                </TopicDescription>
-              )}
-              {editing ? (
-                <>
-                  <br />
-                  <TitleHolder>Author </TitleHolder>
-                  <WrapperInput>
-                    <StyledInput
-                      type="text"
-                      value={newAuthor}
-                      onChange={handleChangeAuthor}
-                    />
-                  </WrapperInput>
-                </>
-              ) : (
-                <>
-                  <Topic>Author: </Topic>
-                  <TopicDescription>{author || 'Author Name'}</TopicDescription>
-                </>
-              )}
-              <>
-                {editing ? (
-                  <WrapperDropDown>
-                    <TitleHolder>Location </TitleHolder>
-                    <RestyledDropdown
-                      options={allLocations!.getAllLocations.map((loc) => ({
-                        id: loc!.id,
-                        value: loc!.location,
-                      }))}
-                      initIndex={currentLocationIndex}
-                      onChange={(option) =>
-                        handleChangeDropdown(option as TUserLocation)
-                      }
-                    />
-                  </WrapperDropDown>
-                ) : (
-                  <>
-                    <Topic>Location: </Topic>
-                    <TopicDescription>{location.location}</TopicDescription>
-                  </>
-                )}
-              </>
-              {userRole === RolesTypes.READER ? (
-                <>
-                  <Topic>State: </Topic>
-                  <StyledStatus>
-                    <BookStatus status={statusInfo?.status} date={created_at} />
-                  </StyledStatus>
-                </>
-              ) : editing ? (
-                <>
-                  <br />
-                  <TitleHolder>Deadline </TitleHolder>
-                  <StyledInputDeadline
-                    value={newDeadline}
-                    type="number"
-                    onChange={handleChangeDeadline}
-                    min="1"
-                    max="31"
-                  />{' '}
-                  days
-                </>
-              ) : (
-                <>
-                  <Topic>Deadline: </Topic>
-                  <TopicDescription>{newDeadline + ' days'}</TopicDescription>
-                </>
-              )}
-            </ShortDescription>
-          </WrapperInfo>
+          <DescriptionBook
+            title={title}
+            author={author}
+            category={category}
+            date={created_at}
+            editing={editing}
+            location={location}
+            src={src}
+            status={statusInfo?.status}
+            newDeadline={newDeadline}
+            newTitleAndAuthor={newDescriptionData}
+            handleChangeDeadline={handleChangeDeadline}
+            handleChangeLocation={handleChangeLocation}
+            handleChangeAuthorAndTitle={handleChangeNewDescriptionData}
+            handleChangeNewGenre={handleChangeCategory}
+          />
           {userRole === RolesTypes.READER ? (
             <>
               {person_id === id ? (
@@ -750,10 +552,7 @@ const BookInfo: FC<IBookInfoProps> = ({
         ) : (
           <LongDescription>
             <Topic>Description: </Topic>
-            <Description>
-              {description ||
-                ' Lorem ipsum dolor sit amet consectetur, adipisicing elit. Sequi impedit aliquid alias consequuntur! Totam sequi expedita sunt dolor obcaecati, iusto ducimus? Beatae ea, commodi ab repellat, corporis atque quasi, tempore sunt modi similique soluta nemo hic necessitatibus esse accusantium omnis neque rerum. Placeat tempore, fugiat unde consequuntur dolor tempora ducimus.'}
-            </Description>
+            <Description>{description || ''}</Description>
             {userRole === RolesTypes.READER ? (
               <OpenLink>see full description</OpenLink>
             ) : null}
